fix cart empty, cart icon, remove wallet, delete product, counter hor

+++ lib/src/models/coupon_list_model.dart
@@ -74,7 +74,7 @@ class Coupons {
   String? title;
   String? code;
   String? discountType;
-  int? discount;
+  double? discount;
   String? image145x110;
   String? endTime;
